Created Hand component to hand toggle functionality
