+++ src/Components/WhiteKard.tsx
@@ -1,56 +1,23 @@
-import React, { useCallback, useContext, useMemo } from "react";
+import React, { useMemo } from "react";
 import { WhiteCard } from "../Types/WhiteCard";
-import { GameContext } from "../State/Game/GameContext";
-import { useHand } from "../State/Hand/HandContext";
-import { SetHandAction } from "../State/Hand/HandActionts";
 import { useUser } from "../State/User/UserContext";
 
 interface WhiteKardProps {
   card: WhiteCard;
+  onClick: (card: WhiteCard) => void;
 }
 
-export const WhiteKard: React.FC<WhiteKardProps> = ({ card }) => {
-  const { blackCard } = useContext(GameContext);
-  const {
-    state: { hand },
-    dispatch,
-  } = useHand();
-
+export const WhiteKard: React.FC<WhiteKardProps> = ({
+  card,
+  onClick = () => {},
+}) => {
   const {
     state: { hasSubmittedCards },
   } = useUser();
 
-  const highestOrder = useMemo(() => {
-    return Math.max(...hand.map((item) => item.order));
-  }, [hand]);
-
   const nonSubmittedCards = useMemo(() => {
     return !card.selected && hasSubmittedCards;
   }, [card, hasSubmittedCards]);
-
-  const toggle = useCallback(() => {
-    if (hasSubmittedCards) return;
-
-    const clone = [...hand];
-    const cardToSelect = clone.find((item) => item.id === card.id);
-
-    if (!cardToSelect) return;
-
-    if (blackCard.pick < highestOrder + 1) {
-      clone.forEach((item) => {
-        if (item.order > 0) {
-          item.order -= 1;
-        }
-        item.selected = item.order !== 0;
-      });
-      cardToSelect.order = highestOrder;
-    } else {
-      cardToSelect.order = highestOrder + 1;
-    }
-    cardToSelect.selected = !card.selected;
-
-    dispatch(new SetHandAction(clone));
-  }, [card, dispatch, hand, hasSubmittedCards, highestOrder, blackCard]);
 
   return (
     <div
@@ -63,7 +30,9 @@ export const WhiteKard: React.FC<WhiteKardProps> = ({ card }) => {
         ${hasSubmittedCards ? "cursor-not-allowed" : ""} 
         ${nonSubmittedCards ? "opacity-25" : ""}
       `}
-      onClick={toggle}
+      onClick={(e) => {
+        onClick(card);
+      }}
       data-testid={`white-card-${card.id}`}
     >
       {card.order > 0 && (
